Update county_rankings.py
This now includes scrapping the long course times aswell. These are then converted to short course times.

This saved csv can then be updated to 'swim_rankings' in the sql database which will then feed through to the county rankings view which has been updated to rank on this basis.